A-CodingStandard: apply style/formatting fixes (done on misnamed branch)
