+++ .gitignore
@@ -18,3 +18,4 @@ text-ui-test/EXPECTED-UNIX.TXT
 
 # User generated local list file
 src/main/java/data/Tkit.txt
+data/Tkit.txt
