+++ src/main/java/tkit/TaskType.java
@@ -1,0 +1,19 @@
+package tkit;
+
+/** Supported task categories. Each value carries a single-letter tag. */
+enum TaskType {
+    TODO("T"),
+    DEADLINE("D"),
+    EVENT("E");
+
+    private final String tag;
+
+    TaskType(String tag) {
+        this.tag = tag;
+    }
+
+    /** Returns the single-letter tag for this type. */
+    public String tag() {
+        return tag;
+    }
+}
